mid way through big refactor removing a lor of ocer sampoling nonsense, dedcorator unit test passes
