+++ autoarray/inversion/pixelization/image_mesh/hilbert.py
@@ -9,7 +9,7 @@ from autoarray.mask.mask_2d import Mask2D
 from autoarray.inversion.pixelization.image_mesh.abstract_weighted import (
     AbstractImageMeshWeighted,
 )
-from autoarray.operators.over_sampling.uniform import OverSamplerUniform
+from autoarray.operators.over_sampling.over_sampler import OverSampler
 from autoarray.inversion.inversion.settings import SettingsInversion
 from autoarray.structures.grids.irregular_2d import Grid2DIrregular
 
@@ -112,7 +112,7 @@ def super_resolution_grid_from(img_2d, mask, mask_radius, pixel_scales, sub_scal
         radius=mask_radius,
     )
 
-    over_sampler = OverSamplerUniform(mask=new_mask, sub_size=sub_scale)
+    over_sampler = OverSampler(mask=new_mask, sub_size=sub_scale)
 
     new_grid = over_sampler.over_sampled_grid
