+++ autoarray/operators/over_sampling/decorator.py
@@ -51,12 +51,10 @@ def over_sample(func):
         if isinstance(grid, Grid2DIrregular) or isinstance(grid, Grid1D):
             return func(obj=obj, grid=grid, *args, **kwargs)
 
-        over_sampled_grid = grid.over_sampler.over_sampled_grid
-
         if obj is not None:
-            values = func(obj, over_sampled_grid, *args, **kwargs)
+            values = func(obj, grid.grid_over_sampled, *args, **kwargs)
         else:
-            values = func(over_sampled_grid, *args, **kwargs)
+            values = func(grid.grid_over_sampled, *args, **kwargs)
 
         return grid.over_sampler.binned_array_2d_from(array=values)
