+++ test_autoarray/dataset/abstract/test_dataset.py
@@ -65,9 +65,7 @@ def test__grid__uses_mask_and_settings(
     masked_imaging_7x7 = ds.AbstractDataset(
         data=masked_image_7x7,
         noise_map=masked_noise_map_7x7,
-        over_sampling=aa.OverSamplingDataset(
-            uniform=aa.OverSamplingUniform(sub_size=2)
-        ),
+        over_sampling=aa.OverSamplingDataset(uniform=aa.OverSampling(sub_size=2)),
     )
 
     assert isinstance(masked_imaging_7x7.grids.uniform, aa.Grid2D)
@@ -97,8 +95,8 @@ def test__grids_pixelization__uses_mask_and_settings(
         data=masked_image_7x7,
         noise_map=masked_noise_map_7x7,
         over_sampling=aa.OverSamplingDataset(
-            uniform=aa.OverSamplingUniform(sub_size=2),
-            pixelization=aa.OverSamplingUniform(sub_size=4),
+            uniform=aa.OverSampling(sub_size=2),
+            pixelization=aa.OverSampling(sub_size=4),
         ),
     )
 
@@ -111,8 +109,8 @@ def test__grid_settings__sub_size(image_7x7, noise_map_7x7):
         data=image_7x7,
         noise_map=noise_map_7x7,
         over_sampling=aa.OverSamplingDataset(
-            uniform=aa.OverSamplingUniform(sub_size=2),
-            pixelization=aa.OverSamplingUniform(sub_size=4),
+            uniform=aa.OverSampling(sub_size=2),
+            pixelization=aa.OverSampling(sub_size=4),
         ),
     )
 
@@ -143,8 +141,8 @@ def test__apply_over_sampling(image_7x7, noise_map_7x7):
         data=image_7x7,
         noise_map=noise_map_7x7,
         over_sampling=aa.OverSamplingDataset(
-            uniform=aa.OverSamplingUniform(sub_size=2),
-            pixelization=aa.OverSamplingUniform(sub_size=2),
+            uniform=aa.OverSampling(sub_size=2),
+            pixelization=aa.OverSampling(sub_size=2),
         ),
     )
 
@@ -162,12 +160,12 @@ def test__apply_over_sampling(image_7x7, noise_map_7x7):
 
     dataset_7x7 = dataset_7x7.apply_over_sampling(
         over_sampling=aa.OverSamplingDataset(
-            uniform=aa.OverSamplingUniform(sub_size=4),
-            pixelization=aa.OverSamplingUniform(sub_size=4),
+            uniform=aa.OverSampling(sub_size=4),
+            pixelization=aa.OverSampling(sub_size=4),
         )
     )
 
-    assert dataset_7x7.over_sampling.uniform.sub_size == 4
+    assert dataset_7x7.over_sampling.over_sampler.sub_size == 4
     assert dataset_7x7.over_sampling.pixelization.sub_size == 4
 
     assert dataset_7x7.grids.uniform[0][0] == pytest.approx(3.0, 1.0e-4)
