+++ test_autoarray/dataset/plot/test_imaging_plotters.py
@@ -20,9 +20,7 @@ def test__individual_attributes_are_output(
     visuals = aplt.Visuals2D(mask=mask_2d_7x7, positions=grid_2d_irregular_7x7_list)
 
     imaging_7x7 = imaging_7x7.apply_over_sampling(
-        over_sampling=aa.OverSamplingDataset(
-            non_uniform=aa.OverSamplingUniform(sub_size=1)
-        )
+        over_sampling=aa.OverSamplingDataset(non_uniform=aa.OverSampling(sub_size=1))
     )
 
     dataset_plotter = aplt.ImagingPlotter(
