+++ test_autoarray/inversion/pixelization/mappers/test_abstract.py
@@ -140,7 +140,7 @@ def test__adaptive_pixel_signals_from___matches_util(grid_2d_7x7, image_7x7):
     )
     pix_weights_for_sub_slim_index = np.ones((9, 1), dtype="int")
 
-    over_sampler = aa.OverSamplerUniform(mask=grid_2d_7x7.mask, sub_size=1)
+    over_sampler = aa.OverSampler(mask=grid_2d_7x7.mask, sub_size=1)
 
     mapper = aa.m.MockMapper(
         source_plane_data_grid=grid_2d_7x7,
@@ -214,7 +214,7 @@ def test__mapped_to_source_from(grid_2d_7x7):
         source_plane_mesh_grid=mesh_grid,
     )
 
-    over_sampler = aa.OverSamplerUniform(mask=grid_2d_7x7.mask, sub_size=1)
+    over_sampler = aa.OverSampler(mask=grid_2d_7x7.mask, sub_size=1)
 
     mapper = aa.Mapper(
         mapper_grids=mapper_grids, over_sampler=over_sampler, regularization=None
