+++ test_autoarray/inversion/pixelization/mappers/test_delaunay.py
@@ -17,7 +17,7 @@ def test__pix_indexes_for_sub_slim_index__matches_util(grid_2d_7x7):
         source_plane_mesh_grid=mesh_grid,
     )
 
-    over_sampler = aa.OverSamplerUniform(mask=grid_2d_7x7.mask, sub_size=1)
+    over_sampler = aa.OverSampler(mask=grid_2d_7x7.mask, sub_size=1)
 
     mapper = aa.Mapper(
         mapper_grids=mapper_grids, over_sampler=over_sampler, regularization=None
