+++ test_autoarray/inversion/pixelization/mappers/test_factory.py
@@ -20,7 +20,7 @@ def test__rectangular_mapper():
     )
 
     # Slightly manipulate input grid so sub gridding is evidence in first source pixel.
-    over_sampler = aa.OverSamplerUniform(mask=mask, sub_size=2)
+    over_sampler = aa.OverSampler(mask=mask, sub_size=2)
     over_sampled_grid = over_sampler.over_sampled_grid
     over_sampled_grid[0, 0] = -2.0
     over_sampled_grid[0, 1] = 2.0
@@ -73,7 +73,7 @@ def test__delaunay_mapper():
     )
 
     # Slightly manipulate input grid so sub gridding is evidence in first source pixel.
-    over_sampler = aa.OverSamplerUniform(mask=mask, sub_size=2)
+    over_sampler = aa.OverSampler(mask=mask, sub_size=2)
     over_sampled_grid = over_sampler.over_sampled_grid
     over_sampled_grid[0, 0] = -2.0
     over_sampled_grid[0, 1] = 2.0
@@ -131,7 +131,7 @@ def test__voronoi_mapper():
     )
 
     # Slightly manipulate input grid so sub gridding is evidence in first source pixel.
-    over_sampler = aa.OverSamplerUniform(mask=mask, sub_size=2)
+    over_sampler = aa.OverSampler(mask=mask, sub_size=2)
     over_sampled_grid = over_sampler.over_sampled_grid
     over_sampled_grid[0, 0] = -2.0
     over_sampled_grid[0, 1] = 2.0
