+++ test_autoarray/inversion/pixelization/mappers/test_voronoi.py
@@ -14,7 +14,7 @@ def test__pix_indexes_for_sub_slim_index__matches_util(grid_2d_7x7):
         values=source_plane_mesh_grid,
     )
 
-    over_sampler = aa.OverSamplerUniform(mask=grid_2d_7x7.mask, sub_size=1)
+    over_sampler = aa.OverSampler(mask=grid_2d_7x7.mask, sub_size=1)
 
     source_plane_mesh_grid = aa.Mesh2DVoronoi(
         values=source_plane_mesh_grid,
