+++ test_autoarray/inversion/pixelization/test_border_relocator.py
@@ -325,7 +325,7 @@ def test__relocated_grid_from__inside_border_no_relocations():
         shape_native=(30, 30), radius=1.0, pixel_scales=(0.1, 0.1)
     )
 
-    over_sampling = aa.OverSamplerUniform(
+    over_sampling = aa.OverSampler(
         mask=mask, sub_size=np.array(mask.pixels_in_mask * [2])
     )
     grid = over_sampling.over_sampled_grid
@@ -345,7 +345,7 @@ def test__relocated_grid_from__outside_border_includes_relocations():
         shape_native=(30, 30), radius=1.0, pixel_scales=(0.1, 0.1)
     )
 
-    over_sampling = aa.OverSamplerUniform(
+    over_sampling = aa.OverSampler(
         mask=mask, sub_size=np.array(mask.pixels_in_mask * [2])
     )
     grid = over_sampling.over_sampled_grid
@@ -368,7 +368,7 @@ def test__relocated_grid_from__positive_origin_included_in_relocate():
         centre=(1.0, 1.0),
     )
 
-    over_sampling = aa.OverSamplerUniform(
+    over_sampling = aa.OverSampler(
         mask=mask, sub_size=np.array(mask.pixels_in_mask * [2])
     )
     grid = over_sampling.over_sampled_grid
