+++ test_autoarray/inversion/test_linear_obj.py
@@ -44,7 +44,7 @@ def test__data_to_pix_unique_from():
 
     mask = aa.Mask2D.all_false(shape_native=(1, 2), pixel_scales=0.1)
 
-    over_sampling = aa.OverSamplerUniform(mask=mask, sub_size=2)
+    over_sampling = aa.OverSampler(mask=mask, sub_size=2)
 
     grid = aa.Grid2D.uniform(
         shape_native=(1, 2), pixel_scales=0.1, over_sampling=over_sampling
