+++ test_autoarray/operators/over_sample/test_uniform.py
@@ -30,7 +30,7 @@ def test__from_sub_size_int():
         pixel_scales=1.0,
     )
 
-    over_sampling = aa.OverSamplerUniform(mask=mask, sub_size=2)
+    over_sampling = aa.OverSampler(mask=mask, sub_size=2)
 
     assert over_sampling.sub_size.slim == pytest.approx([2, 2, 2], 1.0e-4)
     assert over_sampling.sub_size.native == pytest.approx(
@@ -44,7 +44,7 @@ def test__sub_pixel_areas():
         pixel_scales=1.0,
     )
 
-    over_sampling = aa.OverSamplerUniform(mask=mask, sub_size=np.array([1, 2]))
+    over_sampling = aa.OverSampler(mask=mask, sub_size=np.array([1, 2]))
 
     areas = over_sampling.sub_pixel_areas
 
@@ -56,7 +56,7 @@ def test__from_manual_adapt_radial_bin():
 
     grid = aa.Grid2D.from_mask(mask=mask)
 
-    over_sampling = aa.OverSamplingUniform.from_radial_bins(
+    over_sampling = aa.OverSampling.from_radial_bins(
         grid=grid, sub_size_list=[8, 4, 2], radial_list=[1.5, 2.5]
     )
     assert over_sampling.sub_size.native == pytest.approx(
@@ -78,7 +78,7 @@ def test__from_manual_adapt_radial_bin__centre_list_input():
 
     grid = aa.Grid2D.from_mask(mask=mask)
 
-    over_sampling = aa.OverSamplingUniform.from_radial_bins(
+    over_sampling = aa.OverSampling.from_radial_bins(
         grid=grid,
         sub_size_list=[8, 4, 2],
         radial_list=[1.5, 2.5],
@@ -108,7 +108,7 @@ def test__from_adapt():
     data = aa.Array2D(values=[1.0, 2.0, 3.0], mask=mask)
     noise_map = aa.Array2D(values=[1.0, 2.0, 1.0], mask=mask)
 
-    over_sampling = aa.OverSamplingUniform.from_adapt(
+    over_sampling = aa.OverSampling.from_adapt(
         data=data,
         noise_map=noise_map,
         signal_to_noise_cut=1.5,
@@ -118,7 +118,7 @@ def test__from_adapt():
 
     assert over_sampling.sub_size == pytest.approx([2, 2, 4], 1.0e-4)
 
-    over_sampling = aa.OverSamplingUniform.from_adapt(
+    over_sampling = aa.OverSampling.from_adapt(
         data=data,
         noise_map=noise_map,
         signal_to_noise_cut=0.5,
@@ -135,27 +135,11 @@ def test__sub_fraction():
         pixel_scales=1.0,
     )
 
-    over_sampling = aa.OverSamplerUniform(
+    over_sampling = aa.OverSampler(
         mask=mask, sub_size=aa.Array2D(values=[1, 2], mask=mask)
     )
 
     assert over_sampling.sub_fraction.slim == pytest.approx([1.0, 0.25], 1.0e-4)
-
-
-def test__over_sampled_grid():
-    mask = aa.Mask2D(
-        mask=[[False, False], [True, True]],
-        pixel_scales=1.0,
-    )
-
-    over_sampling = aa.OverSamplerUniform(
-        mask=mask, sub_size=aa.Array2D(values=[1, 2], mask=mask)
-    )
-
-    assert over_sampling.over_sampled_grid.native == pytest.approx(
-        np.array([[0.5, -0.5], [0.75, 0.25], [0.75, 0.75], [0.25, 0.25], [0.25, 0.75]]),
-        1.0e-4,
-    )
 
 
 def test__binned_array_2d_from():
@@ -164,7 +148,7 @@ def test__binned_array_2d_from():
         pixel_scales=1.0,
     )
 
-    over_sampling = aa.OverSamplerUniform(
+    over_sampling = aa.OverSampler(
         mask=mask, sub_size=aa.Array2D(values=[1, 2], mask=mask)
     )
 
@@ -182,7 +166,7 @@ def test__sub_mask_index_for_sub_mask_1d_index():
         sub_size=2,
     )
 
-    over_sampling = aa.OverSamplerUniform(mask=mask, sub_size=2)
+    over_sampling = aa.OverSampler(mask=mask, sub_size=2)
 
     sub_mask_index_for_sub_mask_1d_index = (
         aa.util.over_sample.native_sub_index_for_slim_sub_index_2d_from(
@@ -202,7 +186,7 @@ def test__slim_index_for_sub_slim_index():
         sub_size=2,
     )
 
-    over_sampling = aa.OverSamplerUniform(mask=mask, sub_size=2)
+    over_sampling = aa.OverSampler(mask=mask, sub_size=2)
 
     slim_index_for_sub_slim_index_util = (
         aa.util.over_sample.slim_index_for_sub_slim_index_via_mask_2d_from(
